Initial deletion of legacy inspector
