+++ packages/devtools_app/integration_test/test/live_connection/eval_utils.dart
@@ -73,7 +73,7 @@ class EvalTester {
   Future<void> selectWidgetTreeNode(Finder finder) async {
     await tapAndPump(
       find.descendant(
-        of: find.byKey(InspectorScreenBodyState.summaryTreeKey),
+        of: find.byKey(InspectorScreenBodyState.inspectorTreeKey),
         matching: finder,
       ),
     );
