+++ packages/devtools_app/lib/devtools_app.dart
@@ -24,11 +24,11 @@ export 'src/screens/deep_link_validation/deep_links_controller.dart';
 export 'src/screens/deep_link_validation/deep_links_screen.dart';
 export 'src/screens/dtd/dtd_tools_controller.dart';
 export 'src/screens/dtd/dtd_tools_screen.dart';
-export 'src/screens/inspector/inspector_controller.dart';
-export 'src/screens/inspector/inspector_screen_body.dart';
-export 'src/screens/inspector/inspector_tree_controller.dart';
 export 'src/screens/inspector_shared/inspector_screen.dart';
 export 'src/screens/inspector_shared/inspector_screen_controller.dart';
+export 'src/screens/inspector_v2/inspector_controller.dart';
+export 'src/screens/inspector_v2/inspector_screen_body.dart';
+export 'src/screens/inspector_v2/inspector_tree_controller.dart';
 export 'src/screens/logging/log_details_controller.dart';
 export 'src/screens/logging/logging_controller.dart';
 export 'src/screens/logging/logging_screen.dart';
@@ -76,7 +76,7 @@ export 'src/shared/analytics/analytics_controller.dart';
 export 'src/shared/charts/treemap.dart';
 export 'src/shared/console/console_service.dart';
 export 'src/shared/console/eval/eval_service.dart';
-export 'src/shared/console/eval/inspector_tree.dart';
+export 'src/shared/console/eval/inspector_tree_v2.dart';
 export 'src/shared/console/primitives/simple_items.dart';
 export 'src/shared/console/widgets/description.dart';
 export 'src/shared/diagnostics/diagnostics_node.dart';
