+++ packages/devtools_app/lib/src/shared/console/widgets/description.dart
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
-/// @docImport '../../../screens/inspector/inspector_tree_controller.dart';
+/// @docImport '../../../screens/inspector_v2/inspector_tree_controller.dart';
 library;
 
 import 'package:devtools_app_shared/ui.dart';
@@ -17,7 +17,7 @@ import '../../primitives/utils.dart';
 import '../../ui/hover.dart';
 import '../../ui/icons.dart';
 import '../../ui/utils.dart';
-import '../eval/inspector_tree.dart';
+import '../eval/inspector_tree_v2.dart';
 import 'expandable_variable.dart';
 
 final _colorIconMaker = ColorIconMaker();
