+++ packages/devtools_app/lib/src/shared/ui/icons.dart
@@ -13,7 +13,7 @@ library;
 import 'package:devtools_app_shared/ui.dart';
 import 'package:flutter/material.dart';
 
-import '../../screens/inspector/layout_explorer/ui/widgets_theme.dart';
+import '../../screens/inspector_v2/layout_explorer/ui/widgets_theme.dart';
 import 'colors.dart';
 
 class CustomIcon extends StatelessWidget {
