+++ packages/devtools_app/test/screens/inspector_v2/inspector_tree_test.dart
@@ -2,16 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 
-import 'package:devtools_app/devtools_app.dart'
-    hide
-        InspectorController,
-        InspectorTreeController,
-        InspectorTree,
-        InspectorTreeConfig,
-        InspectorTreeNode;
-import 'package:devtools_app/src/screens/inspector_v2/inspector_controller.dart';
+import 'package:devtools_app/devtools_app.dart';
 import 'package:devtools_app/src/screens/inspector_v2/inspector_tree_controller.dart';
-import 'package:devtools_app/src/shared/console/eval/inspector_tree_v2.dart';
 import 'package:devtools_app_shared/ui.dart';
 import 'package:devtools_app_shared/utils.dart';
 import 'package:devtools_test/devtools_test.dart';
